webapi/requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==1.11
djangorestframework==3.8.2
idna==2.7
pkg-resources==0.0.0
python-gnupg==0.4.3
pytz==2018.4
requests==2.19.1
urllib3==1.23
